GitData/requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
elastic-transport==8.1.2
elasticsearch==7.14.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
pydantic==1.9.1
python-gitlab==3.4.0
requests==2.27.1
requests-toolbelt==0.9.1
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn[standard]==0.17.6
aiohttp==3.8.1
strawberry-graphql[fastapi]==0.114.5
locust~=2.10.0